- improvements to mmc GC (still disabled for now as is still slow)   + added roots management to SimCodeC.tpl - make VS happy by   + having an array of 1 records without fields (Unparsing.tpl)   + moving includes around - new file SCodeCheck.mo (more will come here) - better display of elements in SCode.mo - added RCS lines to some files.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@8290 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/meta_modelica_gc_stats.h
@@ -49,9 +49,9 @@ extern "C" {
 /* GC statistics, add more here if needed */
 struct mmc_GC_stats_type
 {
-  long allocated;    /* the total allocated memory */
-  long collected;    /* the total collected memory */
-  long collections;  /* the number of performed collections */
+  size_t allocated;    /* the total allocated memory */
+  size_t collected;    /* the total collected memory */
+  size_t collections;  /* the number of performed collections */
 };
 typedef struct mmc_GC_stats_type mmc_GC_stats_type;
